all server logic online

+++ serverLogic/historyListenerServer.py
@@ -14,6 +14,18 @@ Send a POST request::
 from http.server import BaseHTTPRequestHandler, HTTPServer
 import time
 import json
+import os
+
+def addToFile(filePath, jsonDict):
+    toTruncate = open(filePath, 'rb+')
+    toTruncate.seek(-1, os.SEEK_END)
+    toTruncate.truncate()
+
+    toAppend = open(filePath,'a')
+    toAppend.write(',')
+    json.dump(jsonDict['history'][0], toAppend, ensure_ascii=False)
+    toAppend.write(']');
+
 class S(BaseHTTPRequestHandler):
     def _set_headers(self):
         self.send_response(200)
@@ -33,8 +45,12 @@ class S(BaseHTTPRequestHandler):
         self._set_headers()
 
         jsonDict = json.loads(post_data.decode('utf-8'))
-        toWrite = open('/home/ubuntu/historyStorage/' + jsonDict['name'] + '.json','w')
-        json.dump(jsonDict['history'], toWrite, ensure_ascii=False)
+        filePath = '/home/ubuntu/historyStorage/' + jsonDict['name'] + '.json';
+        if(os.path.isfile(filePath)):
+            addToFile(filePath, jsonDict)
+        else:
+            toWrite = open(filePath, 'w')
+            json.dump(jsonDict['history'], toWrite, ensure_ascii=False)
 
 def run(server_class=HTTPServer, handler_class=S, port=8000):
     server_address = ('0.0.0.0', port)
